// lib/generated/prisma/models/House.ts
export type HouseUncheckedCreateInput = {
  id?: string
  name: string
  description?: Prisma.NullableJsonNullValueInput | runtime.InputJsonValue
  type: $Enums.HouseType
  price: number
  houseArea: number
  plotArea: number
  landCategory: $Enums.LandCategory
  finishing: $Enums.Finishing
  heating: $Enums.Heating
  asphaltToHouse?: boolean
  closedComplex?: boolean
  floor: $Enums.FloorCount
  bedroom: $Enums.BedroomCount
  separatedKitchenLiving?: boolean
  bathroom: $Enums.BathroomCount
  facingMaterial?: $Enums.FacingMaterial | null
  wallMaterial: $Enums.WallMaterial
  insulation?: $Enums.Insulation | null
  hasMinimumDownPayment?: boolean
  houseStatus: $Enums.HouseStatus
  saleStatus: $Enums.SaleStatus
  latitude: number
  longitude: number
  cadastralNumber?: string | null
  yandexDiskLink?: string | null
  isActive?: boolean
  createdAt?: Date | string
  updatedAt?: Date | string
  districtId: string
  developerId: string
  messengers?: Prisma.HouseMessengerUncheckedCreateNestedManyWithoutHouseInput
  phones?: Prisma.HousePhoneUncheckedCreateNestedManyWithoutHouseInput
}